ci: automatically tag version in Github Actions
Version number will increment based on the labels #major, #minor, and #patch included in the pull request. I think the default is #patch if there is no tag.

For this commit, I'm hoping the above will trigger a major version increment, which isn't semantically appropriate but will start us off on a nice 1.0.0